refactor utils module
