+++ pyhodl/apis/models.py
@@ -19,7 +19,7 @@
 """ API client core to fetch data """
 
 from pyhodl.logs import Logger
-from pyhodl.utils.misc import download_with_tor, download
+from pyhodl.utils.network import download, download_with_tor
 
 
 class AbstractApiClient(Logger):
